requirements.txt
flask
pillow
pillow-heif
cloudinary